Implemented division.
We had to handle division by zero, which turns out to be mildly annoying because there are three different responses to dividing by zero. If everything is an int, you get an exception. If you have floats you get `infinity` _unless_ both arguments are 0.0, when you get `NaN`. We dealt with it by first testing the second argument using `zero?` and then ignoring the division operator if that was true.