Move most of core to new interpreter namespace.
We've long known that most of what was in `core` really needed to be in something called `interpreter`, and adding a proper `-main` to `core` forced my hand because of cyclic dependencies. The `gp` namespace depended on `core` (when it really wanted `interpreter` things), but I needed `core` to depend on `gp` so it could create and start up a run.