Introduced `get-score` on individuals.
We had tons of "raw" uses of `(:score individual)` floating around, which made it difficult to change the logic for getting a score. We might want to do this when parallelizing things, however, because scores might become references that are placeholders for computations happening elsewhere in other threads. So I created a `get-score` function and (I think) converted all the "raw" uses of `:score` to use this new function.

And the tests still pass :-)